fix: commit remaining files
