+++ api-gateway/nginx.conf
@@ -11,14 +11,14 @@ http {
 
         location /api/tierlists {
             auth_request /auth;
-            proxy_pass http://be-tierlists/;
+            proxy_pass http://tierlist-service/;
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
         }
 
         location /api/tierlistitems {
-            proxy_pass http://be-tierlistitems/;
+            proxy_pass http://tierlistitem-service/;
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
