+++ backend/tierlistitem-service/Dockerfile
@@ -1,0 +1,4 @@
+FROM openjdk:17
+WORKDIR /app
+COPY ./target/tierlistitem-service-0.0.1-SNAPSHOT.jar /app
+CMD ["java", "-jar", "tierlistitem-service-0.0.1-SNAPSHOT.jar"]
