+++ backend/tierlistitem-service/src/main/java/ch/etmles/tierlistitem-service/Controllers/TierListItemNotFoundAdvice.java
@@ -1,4 +1,4 @@
-package ch.etmles.makelist.Controllers;
+package ch.etmles.tierlistitemservice.Controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
