+++ backend/tierlistitem-service/src/main/java/ch/etmles/tierlistitem-service/Repositories/TierListItemRepository.java
@@ -1,10 +1,10 @@
-package ch.etmles.makelist.Repositories;
+package ch.etmles.tierlistitemservice.Repositories;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import ch.etmles.makelist.Entities.TierListItem;
+import ch.etmles.tierlistitemservice.Entities.TierListItem;
 
 public interface TierListItemRepository extends JpaRepository<TierListItem, Long>{
     List<TierListItem> findByTierListId(Long tierListId);
