+++ backend/tierlistitem-service/src/main/java/ch/etmles/tierlistitem-service/TierlistitemserviceApplication.java
@@ -1,13 +1,13 @@
-package ch.etmles.makelist;
+package ch.etmles.tierlistitemservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MakelistApplication {
+public class TierlistitemserviceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(MakelistApplication.class, args);
+		SpringApplication.run(TierlistitemserviceApplication.class, args);
 	}
 
 }
