+++ backend/tierlistitem-service/src/main/resources/application-dev.properties
