+++ backend/tierlistitem-service/src/main/resources/application-mysql.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/Makelist?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://localhost:3306/Makelist
 spring.datasource.username=root
 spring.datasource.password=1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
