+++ backend/tierlistitem-service/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+spring.application.name=tierlistitem-service
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+server.port=8080
+spring.profiles.active=mysql
