+++ compose.yml
@@ -43,27 +43,41 @@ services:
     volumes:
       - ./api-gateway/nginx.conf:/etc/nginx/nginx.conf:ro
     depends_on:
-      - be-tierlists
-      - be-tierlistitems
+      - tierlist-service
+      - tierlistitem-service
       - keycloak
     networks:
       - public-network
       - sso-network
       - backend-network
-  be-tierlists:
-    hostname: mock-be-tierlists
-    image: ealen/echo-server
+  tierlist-service:
+    build: ./backend/tierlist-service
+    ports:
+      - "8081:8080"
+    depends_on:
+      db-engine:
+        condition: service_healthy
     networks:
       - backend-network
+    environment:
+      SPRING_DATASOURCE_URL: jdbc:mysql://db-engine:3306/Makelist
+      SPRING_DATASOURCE_USERNAME: cpnv
+      SPRING_DATASOURCE_PASSWORD: 1234
+      SPRING_PROFILES_ACTIVE: mysql
+  tierlistitem-service:
+    build: ./backend/tierlistitem-service
+    ports:
+      - "8082:8080"
     depends_on:
-      - db-engine
-  be-tierlistitems:
-    hostname: mock-be-tierlistitems
-    image: ealen/echo-server
+      db-engine:
+        condition: service_healthy
     networks:
       - backend-network
-    depends_on:
-      - db-engine
+    environment:
+      SPRING_DATASOURCE_URL: jdbc:mysql://db-engine:3306/Makelist
+      SPRING_DATASOURCE_USERNAME: cpnv
+      SPRING_DATASOURCE_PASSWORD: 1234
+      SPRING_PROFILES_ACTIVE: mysql
   db-engine:
     image: mysql:9.3
     container_name: db-engine
@@ -80,6 +94,11 @@ services:
       - backend-network
       - db-network
     restart: unless-stopped
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost", "-u", "root", "-p1234"]
+      interval: 30s
+      timeout: 10s
+      retries: 5
 
 volumes:
   db-engine:
