--- specs/Specs-Task003.md
@@ -1,17 +1,0 @@
-# Task 003 - "The employee may not be a minor"
-
-## Objectif
-
-Il vous est demandé, en partant du fichier [classDiagram.md](../docs/classDiagram.puml) livré, d'intégrer la capacité à votre architecture
-de pouvoir gérer une exception si l'employé que l'on tente de recruter et mineur.
-
-Vous remarquerez que le niveau de détail du diagramme s'arrête au niveau des classes. Aucunement besoin de détailler ni
-les attributs, ni les méthodes.
-
-Faites en revanche attention aux classes et à leurs relations !
-
-## Complément
-
-D'un point de vue de l'API, on désire recevoir un message personnalisé ainsi qu'un code d'erreur HTTP adéquat.
-
-Ni le message ni le code ne sont des livrables.
