--- src/main/java/ch/etmles/makelist/Controllers/TierListNotFoundException.java
@@ -1,8 +1,0 @@
-package ch.etmles.makelist.Controllers;
-
-public class TierListNotFoundException extends RuntimeException{
-
-    TierListNotFoundException(Long id){
-        super("Could not find tierlist " + id);
-    }
-}
