--- src/main/java/ch/etmles/makelist/Repositories/TierListRepository.java
@@ -1,9 +1,0 @@
-package ch.etmles.makelist.Repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import ch.etmles.makelist.Entities.TierList;
-
-public interface TierListRepository extends JpaRepository<TierList, Long>{
-
-}
