--- src/main/resources/application.properties
@@ -1,4 +1,0 @@
-spring.application.name=makelist
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.profiles.active=dev
