--- src/test/java/ch/etmles/makelist/MakelistApplicationTests.java
@@ -1,13 +1,0 @@
-package ch.etmles.makelist;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MakelistApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
